'LCS_sequence-to-matrix_CODE' implemented output txts written to. TODO:: Batch execution w/ SequenceList.txt

+++ ForestTools/ForestTools/DistanceMeasureCalculator.h
@@ -21,7 +21,7 @@ namespace distanceMeasure
 		~DistanceMeasureCalculator() = default;
 
 		virtual float operator()(const FileObject& file1, const FileObject& file2) const = 0;
-		virtual float normalize(int differenceCount, long sequencesize) const = 0;
+		virtual float normalize(int differenceCount, int sequencesize) const = 0;
 
 	private:
