+++ ForestTools/ForestTools/LcsDistanceCalculator.h
@@ -21,13 +21,13 @@ namespace distanceMeasure
 		~LcsDistanceCalculator() = default;
 
 		float operator()(const FileObject& file1, const FileObject& file2) const override;
-		float normalize(int differenceCount, long sequencesize) const override;
+		float normalize(int differenceCount, int sequencesize) const override;
 
 	private:
 		//helper funcs
 		const int getArrayIndex(int i, int j, int n) const;
 
-		inline long maxSequenceLength(long sequencesize1, long sequencesize2) const;
+		inline int maxSequenceLength(int sequencesize1, int sequencesize2) const;
 	};
 }
